@@ -1,6 +1,6 @@
 {
   "name": "ffformat",
-  "version": "1.0.6",
+  "version": "1.1.2-4",
   "description": "",
   "main": "lib/index.js",
   "types": "types/index.d.ts",
@@ -19,6 +19,11 @@
     "hooks": {
       "pre-commit": "pretty-quick --pattern='src/*' --staged && npm run lint && npm run test -- --watchAll=false && npm run build",
       "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+    }
+  },
+  "config": {
+    "commitizen": {
+      "path": "cz-conventional-changelog"
     }
   },
   "repository": {
@@ -40,15 +45,18 @@
     "@babel/preset-env": "^7.9.5",
     "@babel/preset-typescript": "^7.9.0",
     "@babel/types": "^7.9.5",
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
     "@types/inquirer": "^6.5.0",
     "@types/jest": "^25.2.1",
-    "@typescript-eslint/eslint-plugin": "^2.28.0",
-    "@typescript-eslint/parser": "^2.28.0",
+    "@types/semver": "^7.3.4",
+    "@typescript-eslint/eslint-plugin": "^4.5.0",
+    "@typescript-eslint/parser": "^4.5.0",
     "babel-jest": "^25.3.0",
-    "eslint": "^6.8.0",
-    "eslint-config-alloy": "^3.7.1",
+    "commitizen": "^4.2.2",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^7.11.0",
+    "eslint-config-alloy": "^3.8.2",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
     "gulp-typescript": "^6.0.0-alpha.1",
@@ -59,9 +67,10 @@
     "pre-commit": "^1.2.2",
     "prettier": "^2.0.4",
     "pretty-quick": "^2.0.1",
+    "semver": "^7.3.2",
     "ts-loader": "^8.0.6",
     "ts-node": "^8.9.0",
-    "typescript": "^3.8.3",
+    "typescript": "^4.0.3",
     "vinyl-buffer": "^1.0.1",
     "webpack": "^5.1.3",
     "webpack-cli": "^4.1.0"
