@@ -1,123 +1,139 @@
-/* eslint-disable @typescript-eslint/no-require-imports */
-const inquirer = require('inquirer');
-const fs = require('fs');
-const path = require('path');
-const packageJson = require('./../package.json');
-const currentVersion = packageJson.version;
-const child_process = require('child_process');
+/* eslint-disable */
+import inquirer from 'inquirer';
+import fs from 'fs';
+import path from 'path';
+import child_process from 'child_process';
+import util from 'util';
+import chalk from 'chalk';
+import semverInc from 'semver/functions/inc';
+import { ReleaseType } from 'semver';
 
-const parseVersion = (versionString: string) => {
-  const [numberVersion, aliasVersionStr] = versionString.split('-');
-  const [large, middle, small] = numberVersion.split('.').map(Number);
+import pkg from '../package.json';
 
-  let aliasSuffix: string | undefined;
-  let aliasVersion: number | undefined;
-  if (aliasVersionStr) {
-    const [suffix, version] = aliasVersionStr.split('.');
-    aliasSuffix = suffix;
-    aliasVersion = Number(version);
-  }
+const exec = util.promisify(child_process.exec);
 
-  return { large, middle, small, aliasSuffix, aliasVersion };
+const run = async (command: string) => {
+  console.log(chalk.green(command));
+  await exec(command);
 };
 
-const currentVersionDetail = parseVersion(currentVersion);
-type NextVersionType = 'large' | 'middle' | 'small' | 'alpha' | 'beta' | 'rc';
+const currentVersion = pkg.version;
 
-const createNextVersion = (
-  versionDetail: ReturnType<typeof parseVersion>,
-  nextType: NextVersionType
-) => {
-  const { large, middle, small, aliasSuffix, aliasVersion } = versionDetail;
-  switch (nextType) {
-    case 'alpha':
-    case 'beta':
-    case 'rc': {
-      return `${large}.${middle}.${small}-${nextType}.${
-        aliasSuffix === nextType ? (aliasVersion || 0) + 1 : 0
-      }`;
-    }
-    case 'large': {
-      return `${large + 1}.${middle}.${small}`;
-    }
-    case 'middle': {
-      return `${large}.${middle + 1}.${small}`;
-    }
-    case 'small': {
-      return `${large}.${middle}.${small + 1}`;
-    }
+const getNextVersions = (): { [key in ReleaseType]: string | null } => ({
+  major: semverInc(currentVersion, 'major'),
+  minor: semverInc(currentVersion, 'minor'),
+  patch: semverInc(currentVersion, 'patch'),
+  premajor: semverInc(currentVersion, 'premajor'),
+  preminor: semverInc(currentVersion, 'preminor'),
+  prepatch: semverInc(currentVersion, 'prepatch'),
+  prerelease: semverInc(currentVersion, 'prerelease'),
+});
+
+const timeLog = (logInfo: string, type: 'start' | 'end') => {
+  let info = '';
+  if (type === 'start') {
+    info = `=> 开始任务：${logInfo}`;
+  } else {
+    info = `✨ 结束任务：${logInfo}`;
   }
-};
-
-const choicesList: NextVersionType[] = ['alpha', 'beta', 'rc', 'large', 'middle', 'small'];
-const CUSTOM_VERSION = '自定义版本';
-
-const timeLog = (logInfo: string) => {
   const nowDate = new Date();
   console.log(
     `[${nowDate.toLocaleString()}.${nowDate
       .getMilliseconds()
       .toString()
-      .padStart(3, '0')}] ${logInfo}`
+      .padStart(3, '0')}] ${info}
+    `,
   );
 };
 
-const release = async (versionStr: string) => {
-  const util = require('util');
-  const exec = util.promisify(child_process.exec);
-
-  // =================== 修改版本 ===================
-  packageJson.version = versionStr;
-  timeLog(`🤔开始修改package.json版本`);
-  await fs.writeFileSync(path.resolve(__dirname, './../package.json'), JSON.stringify(packageJson));
-  await exec("pretty-quick --pattern='package.json'");
-  timeLog('😁修改package.json版本成功');
-
-  // =================== 代码推送git仓库 ===================
-  timeLog('🤔代码开始推送到git仓库');
-  await exec('git add package.json');
-  await exec(`git commit -m "v${versionStr}" -n`);
-  await exec('git push -f');
-  timeLog('😁代码推送到git仓库成功');
-
-  // =================== 打包及发布npm ===================
-  timeLog('🤔开始打包和发布NPM');
-  await exec('npm run build && npm publish');
-  timeLog('😁发布NPM成功');
-
-  // =================== git仓库打TAG ===================
-  timeLog('🤔开始打TAG推送到git仓库');
-  await exec(`git tag v${versionStr}`);
-  await exec(`git push origin tag v${versionStr}`);
-  timeLog('😁打TAG推送到git仓库成功');
-};
-
-let startTime = 0;
-
-inquirer
-  .prompt([
+/**
+ * 询问获取下一次版本号
+ */
+async function prompt(): Promise<string> {
+  const nextVersions = getNextVersions();
+  const { nextVersion } = await inquirer.prompt([
     {
       type: 'list',
-      name: 'version',
-      message: `请选择将要发布的版本 (当前 ${currentVersion})`,
-      choices: choicesList
-        .map((v) => createNextVersion(currentVersionDetail, v))
-        .concat([CUSTOM_VERSION])
-    }
-  ])
-  .then(({ version }: any) => {
-    if (version === CUSTOM_VERSION) {
-      return inquirer.prompt([{ type: 'input', name: 'version', message: '输入自定义版本号' }]);
-    }
-    return { version };
-  })
-  .then(({ version }: any) => {
-    startTime = Date.now();
-    return release(version);
-  })
-  .then(() => {
-    timeLog(`😝发布成功 总共耗时${((Date.now() - startTime) / 1000).toFixed(3)}s`);
-  })
-  .catch((err: any) => {
-    console.log(err, 'error o');
-  });
+      name: 'nextVersion',
+      message: `请选择将要发布的版本 (当前版本 ${currentVersion})`,
+      choices: (Object.keys(nextVersions) as Array<ReleaseType>).map(level => ({
+        name: `${level} => ${nextVersions[level]}`,
+        value: nextVersions[level],
+      })),
+    },
+  ]);
+  return nextVersion;
+}
+
+/**
+ * 更新版本号
+ * @param nextVersion 新版本号
+ */
+async function updateVersion(nextVersion: string) {
+  pkg.version = nextVersion;
+  timeLog('修改package.json版本号', 'start');
+  await fs.writeFileSync(path.resolve(__dirname, './../package.json'), JSON.stringify(pkg));
+  await run('npx prettier package.json --write');
+  await run('git add package.json');
+  timeLog('修改package.json版本号', 'end');
+}
+
+/**
+ * 将代码提交至git
+ */
+async function push(nextVersion: string) {
+  timeLog('推送代码至git仓库', 'start');
+  await run(`git commit -m "v${nextVersion}" -n`);
+  await run('git push');
+  timeLog('推送代码至git仓库', 'end');
+}
+
+/**
+ * 组件库打包
+ */
+async function build() {
+  timeLog('组件库打包', 'start');
+  await run('npm run build');
+  timeLog('组件库打包', 'end');
+}
+
+/**
+ * 发布至npm
+ */
+async function publish() {
+  timeLog('发布组件库', 'start');
+  await run('npm publish');
+  timeLog('发布组件库', 'end');
+}
+
+/**
+ * 打tag提交至git
+ */
+async function tag(nextVersion: string) {
+  timeLog('打tag并推送至git', 'start');
+  await run(`git tag v${nextVersion}`);
+  await run(`git push origin tag v${nextVersion}`);
+  timeLog('打tag并推送至git', 'end');
+}
+
+async function main() {
+  try {
+    const nextVersion = await prompt();
+    const startTime = Date.now();
+    // =================== 更新版本号 ===================
+    await updateVersion(nextVersion);
+    // =================== 代码推送git仓库 ===================
+    await push(nextVersion);
+    // =================== 组件库打包 ===================
+    await build();
+    // =================== 发布至npm ===================
+    await publish();
+    // =================== 打tag并推送至git ===================
+    await tag(nextVersion);
+    console.log(`✨ 发布流程结束 共耗时${((Date.now() - startTime) / 1000).toFixed(3)}s`);
+  } catch (error) {
+    console.log('💣 发布失败，失败原因：', error);
+  }
+}
+
+main();
