@@ -1,1 +1,1 @@
-define((()=>(()=>{"use strict";var e={42:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0});var r=Symbol("array"),n=Symbol("object"),o=Symbol("boolean"),l=Symbol("string"),u=Symbol("number"),a=Symbol("unknown");t.typeCollection={array:r,object:n,string:l,number:u,boolean:o,unknown:a}},185:function(e,t,r){var n=this&&this.__assign||function(){return(n=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e}).apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0});var o=r(42),l=r(629),u=r(246);function a(e){return e instanceof u.default||e instanceof i}var i=function(){function e(e){var t=this,r=e.type,n=e.types,u=e.defaultValue,i=e.systemDefaultValue;if(this.default=function(e){return t.defaultValue=e,t},!n)throw new Error("the constructor need types to initialize");if(l.typeOf(n)!==o.typeCollection.object&&!a(n))throw new Error("types error");this.type=r,this.types=n,this.defaultValue=u,this.systemDefaultValue=i}return Object.defineProperty(e.prototype,"rule",{get:function(){return[this.types,l.computedInitValue(this.defaultValue,this.type,this.systemDefaultValue)]},enumerable:!0,configurable:!0}),Object.defineProperty(e.prototype,"value",{get:function(){var e=this.rule,t=e[0],r=e[1];return[this.type,this.createType(t),r]},enumerable:!0,configurable:!0}),e.prototype.createType=function(e){return a(e)?e.value:Object.entries(e).reduce((function(e,t){var r,o=t[0],l=t[1];if(!a(l))throw new Error("the key: "+o+" 's value need to be extends class 'Typing'");return n(n({},e),((r={})[o]=l.value,r))}),{})},e}();t.default=i},246:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(629),o=function(){function e(e){var t=this,r=e.type,o=e.defaultValue,l=e.systemDefaultValue;this.default=function(e){return t.defaultValue=e,t},this.createValue=function(e,t,r){return[t,n.computedInitValue(e,t,r)]},this.type=r,this.defaultValue=o,this.systemDefaultValue=l}return Object.defineProperty(e.prototype,"value",{get:function(){return this.createValue(this.defaultValue,this.type,this.systemDefaultValue)},enumerable:!0,configurable:!0}),e}();t.default=o},385:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(185),o=r(246),l=r(42),u={string:function(e){return void 0===e&&(e=""),new o.default({defaultValue:e,systemDefaultValue:"",type:l.typeCollection.string})},number:function(e){return void 0===e&&(e=0),new o.default({defaultValue:e,systemDefaultValue:0,type:l.typeCollection.number})},boolean:function(e){return void 0===e&&(e=!1),new o.default({defaultValue:e,systemDefaultValue:!1,type:l.typeCollection.boolean})},array:function(e){return new n.default({type:l.typeCollection.array,types:e,defaultValue:[],systemDefaultValue:[]})},object:function(e){return new n.default({type:l.typeCollection.object,types:e,defaultValue:{},systemDefaultValue:{}})}};t.default=u},39:function(e,t,r){var n=this&&this.__assign||function(){return(n=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e}).apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0});var o=r(42),l=r(629),u=r(690),a=function(e,t){var r=l.typeOf(t)===o.typeCollection.array?t:t.value;switch(r[0]){case o.typeCollection.string:case o.typeCollection.number:case o.typeCollection.boolean:var i=r[0],c=r[1];return u.transformData(e,i,c);case o.typeCollection.array:var s=r,y=(s[0],s[1]),f=(c=s[2],l.typeOf(y)===o.typeCollection.object?o.typeCollection.object:y[0]);if(l.typeOf(e)!==o.typeCollection.array)return c;var p=f===o.typeCollection.object?[o.typeCollection.object,y,a({},[o.typeCollection.object,y,{}])]:y;return e.map((function(e){return a(e,p)}));case o.typeCollection.object:r[0];var d=r[1];return c=r[2],l.typeOf(e)!==o.typeCollection.object?c:Object.entries(d).reduce((function(t,r){var o,l=r[0],u=r[1];return n(n({},t),((o={})[l]=a(e[l],u),o))}),{});default:return null}};t.default=a},690:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(42);t.transformData=function(e,t,r){switch(t){case n.typeCollection.number:var o=Number(null===e?void 0:e);return isNaN(o)?r||0:o;case n.typeCollection.string:return String(e||r||"");case n.typeCollection.boolean:return Boolean(e)}}},465:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(385);t.typeCreator=n.default;var o=r(39);t.format=o.default;var l=r(42);t.typeCollection=l.typeCollection;var u=r(629);t.typeOf=u.typeOf},629:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(42);t.typeOf=function(e){var t={}.toString.call(e).slice(8,-1).toLowerCase();return n.typeCollection[t]?n.typeCollection[t]:n.typeCollection.unknown},t.computedInitValue=function(e,r,n){return t.typeOf(e)===r?e:n},t.isCompositeType=function(e){return[n.typeCollection.array,n.typeCollection.object].includes(e)}}},t={};return function r(n){if(t[n])return t[n].exports;var o=t[n]={exports:{}};return e[n].call(o.exports,o,o.exports,r),o.exports}(465)})()));
+define((()=>(()=>{"use strict";var e={42:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.typeCollection=void 0;var r=Symbol("array"),n=Symbol("object"),o=Symbol("boolean"),u=Symbol("string"),l=Symbol("number"),a=Symbol("unknown");t.typeCollection={array:r,object:n,string:u,number:l,boolean:o,unknown:a}},185:function(e,t,r){var n=this&&this.__assign||function(){return(n=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e}).apply(this,arguments)},o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var u=r(42),l=r(629),a=o(r(246));function i(e){return e instanceof a.default||e instanceof c}var c=function(){function e(e){var t=this,r=e.type,n=e.types,o=e.defaultValue,a=e.systemDefaultValue;if(this.default=function(e){return t.defaultValue=e,t},!n)throw new Error("the constructor need types to initialize");if(l.typeOf(n)!==u.typeCollection.object&&!i(n))throw new Error("types error");this.type=r,this.types=n,this.defaultValue=o,this.systemDefaultValue=a}return Object.defineProperty(e.prototype,"rule",{get:function(){return[this.types,l.computedInitValue(this.defaultValue,this.type,this.systemDefaultValue)]},enumerable:!1,configurable:!0}),Object.defineProperty(e.prototype,"value",{get:function(){var e=this.rule,t=e[0],r=e[1];return[this.type,this.createType(t),r]},enumerable:!1,configurable:!0}),e.prototype.createType=function(e){return i(e)?e.value:Object.entries(e).reduce((function(e,t){var r,o=t[0],u=t[1];if(!i(u))throw new Error("the key: "+o+" 's value need to be extends class 'Typing'");return n(n({},e),((r={})[o]=u.value,r))}),{})},e}();t.default=c},246:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0});var n=r(629),o=function(){function e(e){var t=this,r=e.type,o=e.defaultValue,u=e.systemDefaultValue;this.default=function(e){return t.defaultValue=e,t},this.createValue=function(e,t,r){return[t,n.computedInitValue(e,t,r)]},this.type=r,this.defaultValue=o,this.systemDefaultValue=u}return Object.defineProperty(e.prototype,"value",{get:function(){return this.createValue(this.defaultValue,this.type,this.systemDefaultValue)},enumerable:!1,configurable:!0}),e}();t.default=o},385:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var o=n(r(185)),u=n(r(246)),l=r(42),a={string:function(e){return void 0===e&&(e=""),new u.default({defaultValue:e,systemDefaultValue:"",type:l.typeCollection.string})},number:function(e){return void 0===e&&(e=0),new u.default({defaultValue:e,systemDefaultValue:0,type:l.typeCollection.number})},boolean:function(e){return void 0===e&&(e=!1),new u.default({defaultValue:e,systemDefaultValue:!1,type:l.typeCollection.boolean})},array:function(e){return new o.default({type:l.typeCollection.array,types:e,defaultValue:[],systemDefaultValue:[]})},object:function(e){return new o.default({type:l.typeCollection.object,types:e,defaultValue:{},systemDefaultValue:{}})}};t.default=a},39:function(e,t,r){var n=this&&this.__assign||function(){return(n=Object.assign||function(e){for(var t,r=1,n=arguments.length;r<n;r++)for(var o in t=arguments[r])Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o]);return e}).apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0});var o=r(42),u=r(629),l=r(690),a=function(e,t){var r=u.typeOf(t)===o.typeCollection.array?t:t.value;switch(r[0]){case o.typeCollection.string:case o.typeCollection.number:case o.typeCollection.boolean:var i=r[0],c=r[1];return l.transformData(e,i,c);case o.typeCollection.array:var f=r,s=(f[0],f[1]),y=(c=f[2],u.typeOf(s)===o.typeCollection.object?o.typeCollection.object:s[0]);if(u.typeOf(e)!==o.typeCollection.array)return c;var p=y===o.typeCollection.object?[o.typeCollection.object,s,a({},[o.typeCollection.object,s,{}])]:s;return e.map((function(e){return a(e,p)}));case o.typeCollection.object:r[0];var d=r[1];return c=r[2],u.typeOf(e)!==o.typeCollection.object?c:Object.entries(d).reduce((function(t,r){var o,u=r[0],l=r[1];return n(n({},t),((o={})[u]=a(e[u],l),o))}),{});default:return null}};t.default=a},690:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.transformData=void 0;var n=r(42);t.transformData=function(e,t,r){switch(t){case n.typeCollection.number:var o=Number(null===e?void 0:e);return isNaN(o)?r||0:o;case n.typeCollection.string:return String(e||r||"");case n.typeCollection.boolean:return Boolean(e)}}},465:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.typeOf=t.typeCollection=t.format=t.typeCreator=void 0;var o=n(r(385));t.typeCreator=o.default;var u=n(r(39));t.format=u.default;var l=r(42);Object.defineProperty(t,"typeCollection",{enumerable:!0,get:function(){return l.typeCollection}});var a=r(629);Object.defineProperty(t,"typeOf",{enumerable:!0,get:function(){return a.typeOf}})},629:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.isCompositeType=t.computedInitValue=t.typeOf=void 0;var n=r(42);t.typeOf=function(e){var t={}.toString.call(e).slice(8,-1).toLowerCase();return n.typeCollection[t]?n.typeCollection[t]:n.typeCollection.unknown},t.computedInitValue=function(e,r,n){return t.typeOf(e)===r?e:n},t.isCompositeType=function(e){return[n.typeCollection.array,n.typeCollection.object].includes(e)}}},t={};return function r(n){if(t[n])return t[n].exports;var o=t[n]={exports:{}};return e[n].call(o.exports,o,o.exports,r),o.exports}(465)})()));
