@@ -5,8 +5,8 @@ declare const typeCreator: {
     string: (defaultValue?: string) => PrimitiveType<string>;
     number: (defaultValue?: number) => PrimitiveType<number>;
     boolean: (defaultValue?: boolean) => PrimitiveType<boolean>;
-    array: (types: CompositeType<object> | CompositeType<import("../constant/dataType").AllowedTypes.ArrayType> | PrimitiveType<string> | PrimitiveType<number> | PrimitiveType<boolean> | Record<string, TypeCreator.AllType>) => CompositeType<any[]>;
-    object: (types: Record<string, TypeCreator.AllType>) => CompositeType<object>;
+    array: (types: TypeCreator.ObjectTypeParams | TypeCreator.AllType) => CompositeType<any[]>;
+    object: (types: TypeCreator.ObjectTypeParams) => CompositeType<object>;
 };
 export declare type TypeCreator = typeof typeCreator;
 export default typeCreator;
